Move MockRadarSwiftBridge into its own file
It's now shared by RadarLocationManagerSwiftFailureTests and
RadarLocationManagerSwiftLocationTests in addition to RadarReplayBufferTests,
so it no longer belongs nested in the latter's file.
